Small updates to program admin table.
